pyproject.toml
[project]
name = "data_academy_agentic_day"
version = "0.1.0"
description = "agentic day for the data academy"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.38.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-ai-slim[google]",
]
